+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.10] - 2016-06-29
+### Added
+- `getEventParameter` - returns the event MUI widget parameters for the current widget.  Get the event target, widget name, widget value ( ex: getEventParameter(event, "muiTargetValue") ).  The value is set when creating a widget.  See menu.lua for setting the values and mui.lua callBacks for getting values (example would be actionForSwitch() method).
+- All widgets support a "value" and can be accessed in the callBacks by getEventParameter() method. See mui.lua for examples.
+
 ## [0.1.9] - 2016-06-29
 ### Added
 - "labelFont" option to createToolbar